Wow lots of changes.
Unfortunately there was some losses with the main function
but this is gonna be worth it because I decided to fix
the obj file reader.

This is going to make it a lot better. since things are a little more robust.
im also working on starting testing and whatnot.
This is super beneficial because i want to make sure these things are
going super good.

Besides this, i want to further implement the changes from
engine objects. this is going to remove certain things.

I also will be making specific shapes (ie planes cubes etc.)
just to keep everything a little simpler.

Then were going to move onto the physics system. thats going to be a neat challenge
youre going to have lots of fun with that.

So that means im going to need to make some kind of big descisions on a few things.
I really enjoy this project so far, its such a kind engine to work with.
I love it.